requirements.txt
#django
Django==1.9.4
django-devserver==0.8.0
django-extensions==1.6.7


ipython==5.3.0
requests==2.3.0
Faker==0.7.10
pytz==2016.10
django-geoposition==0.3.0

#DRF
djangorestframework==3.6.2
django-allauth==0.31.0
django-rest-auth==0.9.1
django-rest-knox==3.0.0
django-filter==1.0.1
markdown==2.6.8

#webpack
django-webpack-loader==0.4.1

#phone number fields for django and REST
django-phonenumber-field==1.3.0
phonenumberslite==8.4.3

#team logo
Pillow==4.3.0

#cleanup old files
django-cleanup==1.0.1
